Update model answer menggunakan timestamp

+++ StackExchange-WebService/src/java/stackexchange/webservice/AnswerWS.java
@@ -5,9 +5,12 @@
  */
 package stackexchange.webservice;
 
+import java.sql.Connection;
 import java.sql.PreparedStatement;
 import java.sql.ResultSet;
+import java.sql.Timestamp;
 import java.util.ArrayList;
+import java.util.Calendar;
 import java.util.List;
 import javax.jws.Oneway;
 import javax.jws.WebService;
@@ -16,7 +19,10 @@ import javax.jws.WebParam;
 import javax.jws.WebResult;
 import javax.xml.ws.RequestWrapper;
 import javax.xml.ws.ResponseWrapper;
+import stackexchange.webservice.auth.Auth;
 import stackexchange.webservice.model.Answer;
+import stackexchange.webservice.model.Question;
+import stackexchange.webservice.model.User;
 import stackexchange.webservice.util.Database;
 
 /**
@@ -26,7 +32,7 @@ import stackexchange.webservice.util.Database;
 @WebService(serviceName = "AnswerWS")
 public class AnswerWS {
 
-    /**
+     /**
      * Web service operation
      */
     @WebMethod(operationName = "getAnswers")
@@ -39,7 +45,7 @@ public class AnswerWS {
             PreparedStatement ps = db.getConnection().prepareStatement(sql);
             ResultSet rs = ps.executeQuery();
             while (rs.next()) {
-                Answer answer = new Answer(rs.getInt("id"), rs.getInt("userid"), rs.getInt("questionId"), rs.getString("name"), rs.getString("email"), rs.getString("content"), rs.getDate("dateMade"), "", rs.getInt("vote"));
+                Answer answer = new Answer(rs.getInt("id"), rs.getInt("userid"), rs.getInt("questionId"), rs.getString("name"), rs.getString("email"), rs.getString("content"), rs.getTimestamp("dateMade"), rs.getInt("vote"));
                 answers.add(answer);
             }
             return answers;
@@ -68,7 +74,7 @@ public class AnswerWS {
             PreparedStatement ps = db.getConnection().prepareStatement(sql);
             ResultSet rs = ps.executeQuery();
             while (rs.next()) {
-                Answer answer = new Answer(rs.getInt("id"), rs.getInt("userid"), rs.getInt("questionId"), rs.getString("name"), rs.getString("email"), rs.getString("content"), rs.getDate("dateMade"), "", rs.getInt("vote"));
+                Answer answer = new Answer(rs.getInt("id"), rs.getInt("userid"), rs.getInt("questionId"), rs.getString("name"), rs.getString("email"), rs.getString("content"), rs.getTimestamp("dateMade"), rs.getInt("vote"));
                 answers.add(answer);
             }
             return answers;
@@ -87,25 +93,42 @@ public class AnswerWS {
      */
     @WebMethod(operationName = "addAnswer")
     @Oneway
-    public void addAnswer(@WebParam(name = "answer") Answer answer) {
-        Database db = new Database();
-        try{
-            String values="(";
-            values+= answer.getQuestionId() +",";
-            values+= "'"+ answer.getName() +"',";
-            values+= "'"+ answer.getEmail() +"',";
-            values+= "'"+ answer.getContent() +"',";
-            values+= "'"+ answer.getDateMade() +"',";
-            values+= answer.getVote() +")";
-            String sql="insert into answers values " + values;
-            PreparedStatement ps = db.getConnection().prepareStatement(sql);
-            ps.executeUpdate();
-        }catch(Exception e){
-            
-        }finally{
-            db.closeConnection();
-            db = null;
-        }
+    public void addAnswer(@WebParam(name = "answer") Answer answer, @WebParam(name = "token")String token) {
+    	Auth auth = new Auth();
+    	int ret = auth.check(answer.getEmail(),token);
+        UserWS usrws = new UserWS();
+        User user = usrws.getUserByEmail(answer.getEmail());
+        Timestamp currentTime = new Timestamp(Calendar.getInstance().getTime().getTime());
+        Answer newAnswer = new Answer(user.getId(), answer.getQuestionId(), user.getName(), user.getEmail(), answer.getContent(),currentTime, 0 );
+    	
+    	if(ret == 1 || ret == 0){
+			Database db = new Database();
+	        try{
+	            String values="(";
+                    values+= newAnswer.getUserid() + ",";
+	            values+= newAnswer.getQuestionId() +",";
+	            values+= "'"+ newAnswer.getName() +"',";
+	            values+= "'"+ newAnswer.getEmail() +"',";
+	            values+= "'"+ newAnswer.getContent() +"',";
+	            values+= "'"+ newAnswer.getDateMade() +"',";
+	            values+= newAnswer.getVote() +")";
+	            String sql="insert into answers (userid, questionId, name, email, content,dateMade,vote) values " + values;
+	            Connection conn = db.getConnection();
+                    PreparedStatement ps = conn.prepareStatement(sql);
+	            ps.executeUpdate();
+                    
+                    sql = "update questions set answer= answer+1 where id=" + newAnswer.getQuestionId();
+                    ps = conn.prepareStatement(sql);
+                    ps.executeUpdate();
+                    
+	        }catch(Exception e){
+	            
+	        }finally{
+	            db.closeConnection();
+	            db = null;
+	        }    		
+    	}
+
     }
 
     /**
@@ -113,18 +136,22 @@ public class AnswerWS {
      */
     @WebMethod(operationName = "deleteAnswer")
     @Oneway
-    public void deleteAnswer(@WebParam(name = "id") int id, @WebParam(name = "questionId") int questionId) {
-        Database db = new Database();
-        try{
-            String sql="delete from questions where questionId=" + questionId + "and id=" + id;
-            PreparedStatement ps = db.getConnection().prepareStatement(sql);
-            ps.executeUpdate();
-        }catch(Exception e){
-            
-        }finally{
-            db.closeConnection();
-            //db = null;
-        }
+    public void deleteAnswer(@WebParam(name = "id") int id, @WebParam(name = "questionId") int questionId, @WebParam(name = "email") String email, @WebParam(name = "token") String token) {
+    	Auth auth = new Auth();
+    	int ret = auth.check(email,token);
+    	if(ret == 1 || ret == 0){
+    		Database db = new Database();
+	        try{
+	            String sql="delete from questions where questionId=" + questionId + "and id=" + id;
+	            PreparedStatement ps = db.getConnection().prepareStatement(sql);
+	            ps.executeUpdate();
+	        }catch(Exception e){
+	            
+	        }finally{
+	            db.closeConnection();
+	            //db = null;
+	        }	
+    	}
     }
 
 
@@ -133,28 +160,32 @@ public class AnswerWS {
      */
     @WebMethod(operationName = "voteAnswer")
     @Oneway
-    public void voteAnswer(@WebParam(name = "answer") Answer answer, @WebParam(name = "inc") boolean inc) {
-        Database db = new Database();
-        try{
-            int val=0;
-            if (inc) {
-                val++;
-            }else{
-                val--;
-            }
-            int id = answer.getId();
-            int questionId = answer.getQuestionId();
-            String values="";
-            values+= "vote=val+("+ val + ")";
-            String sql="update answers set " + values + " where id=" + id + "and questionId=" + questionId;
-            PreparedStatement ps = db.getConnection().prepareStatement(sql);
-            ps.executeUpdate();
-        }catch(Exception e){
-            
-        }finally{
-            db.closeConnection();
-            //db = null;
-        }
+    public void voteAnswer(@WebParam(name = "answer") Answer answer, @WebParam(name = "inc") boolean inc, @WebParam(name = "token") String token) {
+    	Auth auth = new Auth();
+    	int ret = auth.check(answer.getEmail(),token);
+    	if(ret == 1 || ret == 0){
+    		Database db = new Database();
+	        try{
+	            int val=0;
+	            if (inc) {
+	                val++;
+	            }else{
+	                val--;
+	            }
+	            int id = answer.getId();
+	            int questionId = answer.getQuestionId();
+	            String values="";
+	            values+= "vote=val+("+ val + ")";
+	            String sql="update answers set " + values + " where id=" + id + "and questionId=" + questionId;
+	            PreparedStatement ps = db.getConnection().prepareStatement(sql);
+	            ps.executeUpdate();
+	        }catch(Exception e){
+	            
+	        }finally{
+	            db.closeConnection();
+	            //db = null;
+	        }	
+    	}
     }
 
     /**
@@ -163,26 +194,30 @@ public class AnswerWS {
     @WebMethod(operationName = "voteAnswer_1")
     @Oneway
     @RequestWrapper(className = "stackexchange.voteAnswer_1")
-    public void voteAnswer(@WebParam(name = "id") int id, @WebParam(name = "questionId") int questionId, @WebParam(name = "inc") boolean inc) {
-        Database db = new Database();
-        try{
-            int val=0;
-            if (inc) {
-                val++;
-            }else{
-                val--;
-            }
-            String values="";
-            values+= "vote=val+("+ val + ")";
-            String sql="update answers set " + values + " where id=" + id + "and questionId=" + questionId;
-            PreparedStatement ps = db.getConnection().prepareStatement(sql);
-            ps.executeUpdate();
-        }catch(Exception e){
-            
-        }finally{
-            db.closeConnection();
-            db = null;
-        }
+    public void voteAnswer(@WebParam(name = "id") int id, @WebParam(name = "questionId") int questionId, @WebParam(name = "inc") boolean inc, @WebParam(name = "email") String email, @WebParam(name = "token") String token) {
+    	Auth auth = new Auth();
+    	int ret = auth.check(email,token);
+    	if(ret == 1 || ret == 0){
+    		Database db = new Database();
+	        try{
+	            int val=0;
+	            if (inc) {
+	                val++;
+	            }else{
+	                val--;
+	            }
+	            String values="";
+	            values+= "vote=val+("+ val + ")";
+	            String sql="update answers set " + values + " where id=" + id + "and questionId=" + questionId;
+	            PreparedStatement ps = db.getConnection().prepareStatement(sql);
+	            ps.executeUpdate();
+	        }catch(Exception e){
+	            
+	        }finally{
+	            db.closeConnection();
+	            db = null;
+	        }	
+    	}
     }
 
     /**
@@ -190,22 +225,27 @@ public class AnswerWS {
      */
     @WebMethod(operationName = "updateAnswer")
     @Oneway
-    public void updateAnswer(@WebParam(name = "answer") Answer answer) {
-        Database db = new Database();
-        try{
-            int id = answer.getId();
-            int questionId = answer.getQuestionId();
-            String values="";
-            values+= "content='"+ answer.getContent() +"',";
-            values+= "dateMade='"+ answer.getDateMade() +"',";
-            String sql="update answers set " + values + " where id=" + id + "and questionId=" + questionId;
-            PreparedStatement ps = db.getConnection().prepareStatement(sql);
-            ps.executeUpdate();
-        }catch(Exception e){
-            
-        }finally{
-            db.closeConnection();
-            //db = null;
-        }
+    public void updateAnswer(@WebParam(name = "answer") Answer answer, @WebParam(name = "token") String token) {
+    	Auth auth = new Auth();
+    	int ret = auth.check(answer.getEmail(),token);
+    	if(ret == 1 || ret == 0){
+    		Database db = new Database();
+	        try{
+	            int id = answer.getId();
+	            int questionId = answer.getQuestionId();
+	            String values="";
+	            values+= "content='"+ answer.getContent() +"',";
+	            values+= "dateMade='"+ answer.getDateMade() +"',";
+	            String sql="update answers set " + values + " where id=" + id + "and questionId=" + questionId;
+	            PreparedStatement ps = db.getConnection().prepareStatement(sql);
+	            ps.executeUpdate();
+	        }catch(Exception e){
+	            
+	        }finally{
+	            db.closeConnection();
+	            //db = null;
+	        }	
+    	}
     }
+
 }
